Refactoring: Moved logic out of `/restart` handler
The handler is now mainly responsible for error handling, logging and
sending correct response to client.

Added an Error type with a Type field to be able to determine what was
the specific problem with handling the request and being able to respond
accordingly (e.g. 4xx vs 5xx). I'm not 100% happy with this approach but
it works and for now will do.