Restart pods by patching their Deployment
Add the `restartedAt` and `restartReason` to
`spec.template.metadata.annotations` will trigger the creation of new pods
with the new specs (because their template changed).

These annotations are also added to the `Deployment` for consistency
and to avoid the hop to the pods if we want to investigate a problem.

Also:
- made go vet happy by documenting exported stuff
- removed unused error types